Fix exhibition view.
Factor out title and languageReviewedText into separate behavior.
Scrap base_exhibition behavior.
Remove languageReviewedText from exhibition metadata fields.

syslabcom/scrum#1199
