+++ src/recensio/plone/profiles/default/types/Review_Article_Journal.xml
@@ -49,6 +49,7 @@
     <element value="recensio.pages_of_review_in_journal" />
     <element value="recensio.printed_review" />
     <element value="recensio.settings_url_in_citation" />
+    <element value="recensio.text_review" />
   </property>
 
   <!-- View information -->
