+++ src/recensio/plone/profiles/default/types/Review_Exhibition.xml
@@ -32,7 +32,7 @@
     <element value="plone.categorization" />
     <element value="plone.locking" />
     <element value="plone.namefromtitle" />
-    <element value="recensio.base_exhibition" />
+    <element value="recensio.base" />
     <element value="recensio.base_review" />
     <element value="recensio.licence" />
     <element value="recensio.settings_url_in_citation" />
