+++ src/recensio/plone/profiles/default/types/Review_Journal.xml
@@ -47,6 +47,7 @@
     <element value="recensio.pages_of_review_in_journal" />
     <element value="recensio.printed_review" />
     <element value="recensio.settings_url_in_citation" />
+    <element value="recensio.text_review" />
   </property>
 
   <!-- View information -->
