+++ src/recensio/plone/profiles/default/types/Review_Monograph.xml
@@ -50,6 +50,7 @@
     <element value="recensio.printed_review" />
     <element value="recensio.serial" />
     <element value="recensio.settings_url_in_citation" />
+    <element value="recensio.text_review" />
   </property>
 
   <!-- View information -->
